update fixing create users and unit test

+++ src/Application/Modules/Users/Commands/DeleteUserCommandHander.cs
@@ -17,7 +17,12 @@ public class DeleteUserCommandHander(
         try
         {
             // Validate if user exists
-            var user = await userRepository.GetUserByIdAsync(new(command.Id), cancellationToken) ?? throw new InvalidOperationException("User not found.");
+            var user = await userRepository.GetUserByIdAsync(new(command.Id), cancellationToken);
+            if (user is null)
+            {
+                logger.LogWarning("User not found with Id: {Id}", command.Id);
+                return Result.Failure(Error.NotFound("UserNotFound", "User not found."));
+            }
 
             // Delete the user
             await userRepository.DeleteUserAsync(new(command.Id), cancellationToken);
